minor fixes
- Fixed dependecies
- Added authoring on top of header and source files

+++ KompactGameEngine/Managers/ShaderManager.cpp
@@ -1,3 +1,10 @@
+//
+//  ShaderMAnager.cpp
+//  KompactGameEngine
+//
+//  Created by Angelo Moro on 06/11/2015
+//
+
 #include "ShaderManager.h"
 using namespace Managers;
 using namespace std;
@@ -44,7 +51,7 @@ GLuint ShaderManager::CreateShader(GLenum shaderType,
 
 	GLuint shader = glCreateShader(shaderType);
 	const char *shader_code_ptr = source.c_str();
-	const int shader_code_size = source.size();
+	const int shader_code_size = (int)source.size();
 
 	glShaderSource(shader, 1, &shader_code_ptr, &shader_code_size);
 	glCompileShader(shader);
